Move getComputerId to the interface
