+++ src/index.ts
@@ -20,3 +20,11 @@ export * from './sync-utils/read-db';
 export * from './sync-utils/write-db';
 export * from './sync-utils/sync-db';
 export * from './sync-utils/sync-device';
+export * from './database-storage/db-storage-interface';
+export * from './conduits/conduit-interface';
+export * from './conduits/download-rsc-conduit';
+export * from './conduits/install-rsc-conduit';
+export * from './conduits/restore-resources-conduit';
+export * from './conduits/sync-databases-conduit';
+export * from './conduits/update-clock-conduit';
+export * from './conduits/update-sync-info-conduit';
